search bar function added
