+++ Frontend_Application/src/components/search-recipes-bar.component.jsx
@@ -1,8 +1,23 @@
 import * as React from "react";
 import TextField from "@mui/material/TextField";
 import Autocomplete from "@mui/material/Autocomplete";
+import axios from "axios";
 
-export default function SearchRecipesBar() {
+export default function SearchRecipesBar(props) {
+  const setSearchedRecipes = props.setSearchedRecipes;
+
+  const handleSearch = (event) => {
+    console.log(event.target.value);
+    axios
+      .get("http://localhost:8080/api/v0/recipes/search/" + event.target.value)
+      .then((response) => {
+        response.data;
+        console.log(response.data);
+      })
+      .then((data) => setSearchedRecipes(data))
+      .catch((error) => console.log(error));
+  };
+
   return (
     <Autocomplete
       id="recipe-search-bar"
@@ -15,6 +30,11 @@ export default function SearchRecipesBar() {
           InputProps={{
             ...params.InputProps,
             type: "search",
+            onKeyPress: (event) => {
+              if (event.key === "Enter") {
+                handleSearch(event);
+              }
+            },
           }}
         />
       )}
