+++ Frontend_Application/src/routes/dashboard/dashboard.component.jsx
@@ -6,6 +6,7 @@ import Typography from "@mui/material/Typography";
 import Menu from "../../components/menu.component.jsx";
 import SearchRecipesBar from "../../components/search-recipes-bar.component.jsx";
 import useAuth from "../../auth-context/useAuth.jsx";
+import MyRecipesCardContainer from "../../components/MyRecipesCardContainer.component.jsx";
 
 function Dashboard(props) {
   const { user } = useAuth();
@@ -13,6 +14,7 @@ function Dashboard(props) {
   const location = useLocation(); // Use useLocation hook to access location
   const [userData, setUserData] = useState(null);
   const userInfo = props.userInfo;
+  const [searchedRecipes, setSearchedRecipes] = useState([]);
 
   console.log(userInfo, "dash");
   useEffect(() => {
@@ -51,7 +53,8 @@ function Dashboard(props) {
           <Typography variant="h3" gutterBottom sx={{ p: 4 }}>
             Welcome, {userData ? userData.firstName : userInfo.email} to Your Dashboard
           </Typography>
-          <SearchRecipesBar />
+          <SearchRecipesBar setSearchedRecipes={setSearchedRecipes} />
+          <MyRecipesCardContainer recipes={searchedRecipes} />
         </Grid>
         <Grid item>
           <Menu />
